+++ src/physics.cpp
@@ -6,11 +6,11 @@ using namespace physics;
 
 void physics::update(rigidbody* rb)
 {
-	rb->velocity += rb->force * engine::delta_time / rb->mass;
-	rb->position += rb->velocity * engine::delta_time;
+	rb->velocity += rb->force * (float)engine::delta_time / rb->mass;
+	rb->position += rb->velocity * (float)engine::delta_time;
 	rb->transform = glm::translate(glm::mat4x4(1.0f), rb->position);
 
-	if (glm::length(rb->torque) != 0.0f) rb->angular_speed = glm::rotate(rb->angular_speed, glm::length(rb->torque) * engine::delta_time / rb->moment_of_inertia, glm::normalize(rb->torque));
+	if (glm::length(rb->torque) != 0.0f) rb->angular_speed = glm::rotate(rb->angular_speed, glm::length(rb->torque) * (float)engine::delta_time / rb->moment_of_inertia, glm::normalize(rb->torque));
 	rb->rotation *= rb->angular_speed;
 	rb->transform *= rb->rotation;
 }
